Scatterbrained BSP loading work - Moved parts of bsp loading to their own modules - Embedded textures loading is working with the default quake palette
