+++ Cargo.toml
@@ -35,6 +35,7 @@ toml = "0.8"
 once_cell = "1.19.0"
 disjoint-sets = "0.4.2"
 exoquant = "0.2.0"
+rayon = "1.10.0"
 # ogre = "0.1.7"
 
 [dev-dependencies.bevy]
