+++ src/loader/bsp/quake_palette.txt
@@ -1,0 +1,32 @@
+000000,0f0f0f,1f1f1f,2f2f2f,3f3f3f,4b4b4b,5b5b5b,6b6b6b,
+7b7b7b,8b8b8b,9b9b9b,ababab,bbbbbb,cbcbcb,dbdbdb,ebebeb,
+0f0b07,170f0b,1f170b,271b0f,2f2313,372b17,3f2f17,4b371b,
+533b1b,5b431f,634b1f,6b531f,73571f,7b5f23,836723,8f6f23,
+0b0b0f,13131b,1b1b27,272733,2f2f3f,37374b,3f3f57,474767,
+4f4f73,5b5b7f,63638b,6b6b97,7373a3,7b7baf,8383bb,8b8bcb,
+000000,070700,0b0b00,131300,1b1b00,232300,2b2b07,2f2f07,
+373707,3f3f07,474707,4b4b0b,53530b,5b5b0b,63630b,6b6b0f,
+070000,0f0000,170000,1f0000,270000,2f0000,370000,3f0000,
+470000,4f0000,570000,5f0000,670000,6f0000,770000,7f0000,
+131300,1b1b00,232300,2f2b00,372f00,433700,4b3b07,574307,
+5f4707,6b4b0b,77530f,835713,8b5b13,975f1b,a3631f,af6723,
+231307,2f170b,3b1f0f,4b2313,572b17,632f1f,733723,7f3b2b,
+8f4333,9f4f33,af632f,bf772f,cf8f2b,dfab27,efcb1f,fff31b,
+0b0700,1b1300,2b230f,372b13,47331b,533723,633f2b,6f4733,
+7f533f,8b5f47,9b6b53,a77b5f,b7876b,c3937b,d3a38b,e3b397,
+ab8ba3,9f7f97,937387,8b677b,7f5b6f,775363,6b4b57,5f3f4b,
+573743,4b2f37,43272f,371f23,2b171b,231313,170b0b,0f0707,
+bb739f,af6b8f,a35f83,975777,8b4f6b,7f4b5f,734353,6b3b4b,
+5f333f,532b37,47232b,3b1f23,2f171b,231313,170b0b,0f0707,
+dbc3bb,cbb3a7,bfa39b,af978b,a3877b,977b6f,876f5f,7b6353,
+6b5747,5f4b3b,533f33,433327,372b1f,271f17,1b130f,0f0b07,
+6f837b,677b6f,5f7367,576b5f,4f6357,475b4f,3f5347,374b3f,
+2f4337,2b3b2f,233327,1f2b1f,172317,0f1b13,0b130b,070b07,
+fff31b,efdf17,dbcb13,cbb70f,bba70f,ab970b,9b8307,8b7307,
+7b6307,6b5300,5b4700,4b3700,3b2b00,2b1f00,1b0f00,0b0700,
+0000ff,0b0bef,1313df,1b1bcf,2323bf,2b2baf,2f2f9f,2f2f8f,
+2f2f7f,2f2f6f,2f2f5f,2b2b4f,23233f,1b1b2f,13131f,0b0b0f,
+2b0000,3b0000,4b0700,5f0700,6f0f00,7f1707,931f07,a3270b,
+b7330f,c34b1b,cf632b,db7f3b,e3974f,e7ab5f,efbf77,f7d38b,
+a77b3b,b79b37,c7c337,e7e357,7fbfff,abe7ff,d7ffff,670000,
+8b0000,b30000,d70000,ff0000,fff393,fff7c7,ffffff,9f5b53
