+++ src/util.rs
@@ -295,3 +295,13 @@ pub(crate) fn trenchbroom_gltf_rotation_fix(world: &mut World, entity: Entity) {
         }
     }
 }
+
+pub(crate) fn invalid_data(err: impl std::error::Error + Send + Sync + 'static) -> io::Error {
+    io::Error::new(io::ErrorKind::InvalidData, err)
+}
+
+/// Internal bevy_trenchbroom macro that adds an extra error message to io errors.
+#[macro_export]
+macro_rules! add_msg {($($args:tt)+) => {
+    move |err| io::Error::new(err.kind(), format!("{}: {}", format!($($args)+), err.into_inner().map(|err| err.to_string()).unwrap_or_default()))
+};}
